Extract verifyChecksumMatch into mxscopy module
