+++ online_nearline/src/main/scala/FileCopier.scala
@@ -1,5 +1,6 @@
 import akka.stream.Materializer
 import akka.stream.scaladsl.FileIO
+import com.gu.multimedia.mxscopy.ChecksumChecker
 import com.gu.multimedia.mxscopy.helpers.{Copier, MatrixStoreHelper, MetadataHelper}
 import com.gu.multimedia.mxscopy.models.ObjectMatrixEntry
 import com.gu.multimedia.mxscopy.streamcomponents.ChecksumSink
@@ -11,7 +12,7 @@ import java.util.Map
 import scala.concurrent.{ExecutionContext, Future}
 import scala.util.{Success, Try}
 
-class FileCopier()(implicit ec:ExecutionContext, mat:Materializer) {
+class FileCopier(checksumHelper: ChecksumChecker)(implicit ec:ExecutionContext, mat:Materializer) {
   private val logger = LoggerFactory.getLogger(getClass)
 
   private val numberedMatcherWithExt = "-(\\d+)\\.[^.]+$".r.unanchored
@@ -135,34 +136,9 @@ class FileCopier()(implicit ec:ExecutionContext, mat:Materializer) {
    * @param maybeLocalChecksum stored local checksum; if set this is used instead of re-calculating. Leave this out when calling.
    * @return a Future containing the OID of the first matching file if present or None otherwise
    */
-  protected def verifyChecksumMatch(filePath:Path, potentialFiles:Seq[MxsObject], maybeLocalChecksum:Option[String]=None):Future[Option[String]] = potentialFiles.headOption match {
-    case None=>
-      logger.info(s"$filePath: No matches found for file checksum")
-      Future(None)
-    case Some(mxsFileToCheck)=>
-      logger.info(s"$filePath: Verifying checksum for MXS file ${mxsFileToCheck.getId}")
-      val savedContext = getContextMap()  //need to save the debug context for when we go in and out of akka
-      val requiredChecksums = Seq(
-        maybeLocalChecksum.map(cs=>Future(Some(cs))).getOrElse(getChecksumFromPath(filePath)),
-        getOMFileMd5(mxsFileToCheck)
-      )
-      Future
-        .sequence(requiredChecksums)
-        .map(results => {
-          if(savedContext.isDefined) setContextMap(savedContext.get)
-          val fileChecksum      = results.head.asInstanceOf[Option[String]]
-          val applianceChecksum = results(1).asInstanceOf[Try[String]]
-          logger.info(s"$filePath: local checksum is $fileChecksum, ${mxsFileToCheck.getId} checksum is $applianceChecksum")
-          (fileChecksum == applianceChecksum.toOption, fileChecksum)
-        })
-        .flatMap({
-          case (true, _)=>
-            logger.info(s"$filePath: Got a checksum match for remote file ${mxsFileToCheck.getId}")
-            Future(Some(potentialFiles.head.getId))  //true => we got a match
-          case (false, localChecksum)=>
-            logger.info(s"$filePath: ${mxsFileToCheck.getId} did not match, trying the next entry of ${potentialFiles.tail.length}")
-            verifyChecksumMatch(filePath, potentialFiles.tail, localChecksum)
-        })
+
+  protected def verifyChecksumMatchUsingHelper(filePath:Path, potentialFiles:Seq[MxsObject], maybeLocalChecksum:Option[String]=None):Future[Option[String]] =  {
+    checksumHelper.verifyChecksumMatch(filePath, potentialFiles, maybeLocalChecksum)
   }
 
   protected def openMxsObject(vault:Vault, oid:String) = Try { vault.getObject(oid) }
@@ -182,7 +158,7 @@ class FileCopier()(implicit ec:ExecutionContext, mat:Materializer) {
       fileSize <- Future.fromTry(Try { getSizeFromPath(filePath) })
       potentialMatches <- findMatchingFilesOnNearline(vault, filePath, fileSize)
       potentialMatchesFiles <- Future.sequence(potentialMatches.map(entry=>Future.fromTry(openMxsObject(vault, entry.oid))))
-      alreadyExists <- verifyChecksumMatch(filePath, potentialMatchesFiles)
+      alreadyExists <- verifyChecksumMatchUsingHelper(filePath, potentialMatchesFiles)
       result <- alreadyExists match {
         case Some(existingId)=>
           logger.info(s"$filePath: Object already exists with object id ${existingId}")
